+++ UI/README.txt
@@ -1,0 +1,1 @@
+Здесь будут все экраны и менюшки, а так же сам файл с классом/классами для UI (например, настройки, главный экран и т.д.)
